requirements_dev.txt
#---------------------------------------------------------
# PlexAPI requirements to run py.test.
# pip install -r requirements_dev.txt
#---------------------------------------------------------
coveralls==3.3.1;python_version>="3.8"
flake8==6.0.0;python_version>="3.8"
pillow==10.0.0;python_version>="3.8"
pytest==7.4.0;python_version>="3.8"
pytest-cache==1.0;python_version>="3.8"
pytest-cov==4.1.0;python_version>="3.8"
pytest-mock==3.11.1;python_version>="3.8"
recommonmark==0.7.1;python_version>="3.8"
requests-mock==1.11.0;python_version>="3.8"
sphinx==6.2.1;python_version>="3.8"
sphinx-rtd-theme==1.2.2;python_version>="3.8"
tqdm==4.65.0;python_version>="3.8"
websocket-client==1.6.1;python_version>="3.8"

-r requirements.txt
# python 2.7
# everything is specifically marked as python_version<"3" because we merge it to the existing requirements_dev.txt
# these are for backporting the project, not needed for tests
autopep8==1.5.7;python_version<"3"
future-fstrings==1.2.0;python_version<"3"

# this is needed for linting tests only
flake8==3.9.2;python_version<"3"

# these are needed for tests
coveralls==1.11.1;python_version<"3"
pillow==6.2.2;python_version<"3"
pytest==4.6.11;python_version<"3"
pytest-cache==1.0;python_version<"3"
pytest-cov==2.12.1;python_version<"3"
pytest-mock==2.0.0;python_version<"3"
recommonmark==0.7.1;python_version<"3"
requests-mock==1.11.0;python_version<"3"
sphinx==1.8.6;python_version<"3"
sphinx-rtd-theme==0.5.2;python_version<"3"
tqdm==4.64.1;python_version<"3"
websocket-client==0.59.0;python_version<"3"

# python > 3 and < 3.8
# for tests only
coveralls>=1.11.1;python_version>="3" and python_version<"3.8"
pillow>=6.2.2;python_version>="3" and python_version<"3.8"
pytest>=4.6.11;python_version>="3" and python_version<"3.8"
pytest-cache>=1.0;python_version>="3" and python_version<"3.8"
pytest-cov>=2.12.1;python_version>="3" and python_version<"3.8"
pytest-mock>=2.0.0;python_version>="3" and python_version<"3.8"
recommonmark>=0.7.1;python_version>="3" and python_version<"3.8"
requests-mock>=1.11.0;python_version>="3" and python_version<"3.8"
sphinx>=1.8.6;python_version>="3" and python_version<"3.8"
sphinx-rtd-theme>=0.5.2;python_version>="3" and python_version<"3.8"
tqdm>=4.64.1;python_version>="3" and python_version<"3.8"
websocket-client>=0.59.0;python_version>="3" and python_version<"3.8"
